+++ server/package.json
@@ -10,7 +10,9 @@
     "author": "delesslin",
     "license": "ISC",
     "dependencies": {
+        "dotenv": "^16.5.0",
         "express": "^5.1.0",
+        "mongoose": "^8.15.1",
         "showdown": "^2.1.0"
     }
 }
